+++ ObjectPrinting.Tests/ObjectPrinterAcceptanceTests.cs
@@ -1,0 +1,16 @@
+﻿using NUnit.Framework;
+
+namespace ObjectPrinting.Tests
+{
+    [TestFixture]
+    public class ObjectPrinterAcceptanceTests
+    {
+        [Test]
+        public void Demo()
+        {
+            var person = new Person { Name = "Alex", Age = 19 };
+            var printer = ObjectPrinter.For<Person>();
+            printer.PrintToString(person);
+        }
+    }
+}
