--- ObjectPrinting/Solved/PropertyPrintingConfig.cs
@@ -1,32 +1,0 @@
-﻿using System;
-using System.Globalization;
-
-namespace ObjectPrinting.Solved
-{
-    public class PropertyPrintingConfig<TOwner, TPropType> : IPropertyPrintingConfig<TOwner, TPropType>
-    {
-        private readonly PrintingConfig<TOwner> printingConfig;
-
-        public PropertyPrintingConfig(PrintingConfig<TOwner> printingConfig)
-        {
-            this.printingConfig = printingConfig;
-        }
-
-        public PrintingConfig<TOwner> Using(Func<TPropType, string> print)
-        {
-            return printingConfig;
-        }
-
-        public PrintingConfig<TOwner> Using(CultureInfo culture)
-        {
-            return printingConfig;
-        }
-
-        PrintingConfig<TOwner> IPropertyPrintingConfig<TOwner, TPropType>.ParentConfig => printingConfig;
-    }
-
-    public interface IPropertyPrintingConfig<TOwner, TPropType>
-    {
-        PrintingConfig<TOwner> ParentConfig { get; }
-    }
-}
